split Kontena::Cli::Certificate::Common#show_certificate
